+++ seqan/projects/library/apps/seqcons/seqcons.cpp
@@ -54,24 +54,24 @@ int main(int argc, const char *argv[]) {
 	addSection(parser, "Main Options:");
 	addOption(parser, addArgumentText(CommandLineOption("r", "reads", "file with reads", OptionType::String), "<FASTA reads file>"));
 	addOption(parser, addArgumentText(CommandLineOption("a", "afg", "message file", OptionType::String), "<AMOS afg file>"));
-	addOption(parser, addArgumentText(CommandLineOption("o", "outfile", "output filename", OptionType::String, "align.txt"), "<Filename>"));
-	addOption(parser, addArgumentText(CommandLineOption("f", "format", "output format", OptionType::String, "afg"), "[seqan | afg]"));
-	addOption(parser, addArgumentText(CommandLineOption("m", "method", "alignment method", OptionType::String, "realign"), "[realign | msa]"));
-	addOption(parser, addArgumentText(CommandLineOption("b", "bandwidth", "bandwidth", OptionType::Int, 8), "<Int>"));
+	addOption(parser, addArgumentText(CommandLineOption("o", "outfile", "output filename", (int)OptionType::String, "align.txt"), "<Filename>"));
+	addOption(parser, addArgumentText(CommandLineOption("f", "format", "output format", (int)OptionType::String, "afg"), "[seqan | afg]"));
+	addOption(parser, addArgumentText(CommandLineOption("m", "method", "alignment method", (int)OptionType::String, "realign"), "[realign | msa]"));
+	addOption(parser, addArgumentText(CommandLineOption("b", "bandwidth", "bandwidth", (int)OptionType::Int, 8), "<Int>"));
 	addOption(parser, CommandLineOption("n", "noalign", "no align, only convert input", OptionType::Boolean));
 
 	addSection(parser, "MSA Method Options:");
-	addOption(parser, addArgumentText(CommandLineOption("ma", "matchlength", "min. overlap length", OptionType::Int, 15), "<Int>"));
-	addOption(parser, addArgumentText(CommandLineOption("qu", "quality", "min. overlap precent identity", OptionType::Int, 80), "<Int>"));
-	addOption(parser, addArgumentText(CommandLineOption("ov", "overlaps", "min. number of overlaps per read", OptionType::Int, 3), "<Int>"));
-	addOption(parser, addArgumentText(CommandLineOption("wi", "window", "window size", OptionType::Int, 0), "<Int>"));
+	addOption(parser, addArgumentText(CommandLineOption("ma", "matchlength", "min. overlap length", (int)OptionType::Int, 15), "<Int>"));
+	addOption(parser, addArgumentText(CommandLineOption("qu", "quality", "min. overlap precent identity", (int)OptionType::Int, 80), "<Int>"));
+	addOption(parser, addArgumentText(CommandLineOption("ov", "overlaps", "min. number of overlaps per read", (int)OptionType::Int, 3), "<Int>"));
+	addOption(parser, addArgumentText(CommandLineOption("wi", "window", "window size", (int)OptionType::Int, 0), "<Int>"));
 	addHelpLine(parser, "/*If this parameter is > 0 then all");
 	addHelpLine(parser, "  overlaps within a given window");
 	addHelpLine(parser, "  are computed.*/");
 
 	addSection(parser, "ReAlign Method Options:");
 	addOption(parser, CommandLineOption("in", "include", "include contig sequence", OptionType::Boolean));
-	addOption(parser, addArgumentText(CommandLineOption("rm", "rmethod", "realign method", OptionType::String, "gotoh"), "[nw | gotoh]"));
+	addOption(parser, addArgumentText(CommandLineOption("rm", "rmethod", "realign method", (int)OptionType::String, "gotoh"), "[nw | gotoh]"));
 
 	if (argc == 1)
 	{
